requirements.txt
flask
gspread
google-auth
python-dotenv
flask-cors
gunicorn
